+++ A32NX/html_ui/Pages/A32NX_Core/FMGC/A32NX_ConstraintManager.js
@@ -7,10 +7,10 @@ class A32NX_ConstraintManager {
      * @returns {number} valid constraint altitude
      */
     static getDisplayedConstraintAltitude(flightPhase, fcuSelAlt, constraintAlt) {
-        if (!constraintAlt || flightPhase === FlightPhase.FLIGHT_PHASE_CRUISE) {
+        if (!constraintAlt || flightPhase === FmgcFlightPhases.CRUISE) {
             return 0;
         }
-        if (flightPhase === FlightPhase.FLIGHT_PHASE_DESCENT || flightPhase === FlightPhase.FLIGHT_PHASE_APPROACH) {
+        if (flightPhase === FmgcFlightPhases.DESCENT || flightPhase === FmgcFlightPhases.APPROACH) {
             return fcuSelAlt > constraintAlt ? 0 : constraintAlt;
         }
         return fcuSelAlt < constraintAlt ? 0 : constraintAlt;
@@ -25,10 +25,10 @@ class A32NX_ConstraintManager {
      * @returns {number} new cached altitude constraint
      */
     static getConstraintAltitude(flightPhase, fpm, curConstraintAlt, crzAlt) {
-        if (fpm.getIsDirectTo() || flightPhase === FlightPhase.FLIGHT_PHASE_CRUISE) {
+        if (fpm.getIsDirectTo() || flightPhase === FmgcFlightPhases.CRUISE) {
             return 0;
         }
-        if (flightPhase === FlightPhase.FLIGHT_PHASE_DESCENT || flightPhase === FlightPhase.FLIGHT_PHASE_APPROACH) {
+        if (flightPhase === FmgcFlightPhases.DESCENT || flightPhase === FmgcFlightPhases.APPROACH) {
             return Math.round(this._getAltitudeConstraintDescent(fpm, curConstraintAlt));
         }
         return Math.round(this._getAltitudeConstraintAscent(fpm, crzAlt));